extracted useEffect into fetchResponsePageData
